requirements.txt
pyside2>=5.15
mysql-connector-python==8.0.23
configparser==5.0.2